// src/app/models/procedureparam.ts
export class ReservaParams {
  pCampo0: string = 'idreservas';
  pValor0: string = '';
  pCampo1: string = 'nombrereserva';
  pValor1: string = '';
  pCampo2: string = 'fechareserva';
  pValor2: string = '';
  pCampo3: string = 'horareserva';
  pValor3: string = '';
  pCampo4: string = 'numpersonas';
  pValor4: string = '';
  pCampo5: string = 'estado';
  pValor5: string = '';
  pCampo6: string = 'fecharegistro';
  pValor6: string = '';
  pCampo7: string = 'estadoregistro';
  pValor7: string = '';

  public inicia() {
      this.pCampo0 = 'idreservas';
      this.pValor0 = '';
      this.pCampo1 = 'nombrereserva';
      this.pValor1 = '';
      this.pCampo2 = 'fechareserva';
      this.pValor2 = '';
      this.pCampo3 = 'horareserva';
      this.pValor3 = '';
      this.pCampo4 = 'numpersonas';
      this.pValor4 = '';
      this.pCampo5 = 'estado';
      this.pValor5 = '';
      this.pCampo6 = 'fecharegistro';
      this.pValor6 = '';
      this.pCampo7 = 'estadoregistro';
      this.pValor7 = '';
  }
}